MB-34790 - add CSV option to 'cbsummary' tool
'cbsummary' now has a '--csv' option to produce the report in CSV
format.

This change also fixes a bug where in the default report, the total
number of nodes and versions was incorrect.

Finally, since information about the number of 'cores' is not
available prior to release 6.5, for earlier versions the number of
cores is listed as 'N/A' in the CSV report.

Change-Id: Ib7ff9a8676170a9fd975f9c6ec45b6ff6ee178dd
Reviewed-on: http://review.couchbase.org/111550
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>